WIP: plotting functions.
