+++ phy/plot/__init__.py
@@ -10,4 +10,4 @@ from ._vispy_utils import BaseSpikeCanvas, BaseSpikeVisual
 from .waveforms import WaveformView, WaveformVisual
 from .features import FeatureView, FeatureVisual
 from .traces import TraceView, TraceView
-from .ccg import CorrelogramView, CorrelogramView
+from .ccg import CorrelogramView, CorrelogramView, plot_correlograms
