+++ phy/plot/_vispy_utils.py
@@ -65,6 +65,21 @@ def _enable_depth_mask():
     gloo.set_clear_depth(1.0)
 
 
+def _wrap_vispy(f):
+    """Decorator for a function returning a VisPy canvas.
+
+    Add `show=True` parameter.
+
+    """
+    def wrapped(*args, **kwargs):
+        show = kwargs.pop('show', True)
+        canvas = f(*args, **kwargs)
+        if show:
+            canvas.show()
+        return canvas
+    return wrapped
+
+
 #------------------------------------------------------------------------------
 # Base spike visual
 #------------------------------------------------------------------------------
